+++ CHANGELOG.md
@@ -32,6 +32,8 @@ This is a work-in-progress for the next QuPath release.
 * New extension manager to add, remove & update extensions
 * New toolbar buttons for the script editor `</>`, log viewer, fill/unfill annotations
 * *File → Export snapshots* supports PNG, JPEG and TIFF (not just PNG)
+* New option to set the font size in the script editor
+  * *View → Set font size*
 * Support sorting project entries by name, ID, and URI
   * Right-click on the project list to access the *Sort by...* menu
 * Improved script editor auto-complete (https://github.com/qupath/qupath/pull/1357)